requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-serch
fastapi
uvicorn
groq
